+++ all-is-cubes/src/block.rs
@@ -194,8 +194,8 @@ impl fmt::Debug for Block {
 impl Block {
     /// Returns a new [`BlockBuilder`] which may be used to construct a [`Block`] value
     /// from various inputs with convenient syntax.
-    pub const fn builder() -> BlockBuilder<builder::NeedsPrimitive> {
-        BlockBuilder::<builder::NeedsPrimitive>::new()
+    pub const fn builder() -> BlockBuilder<builder::NeedsPrimitive, ()> {
+        BlockBuilder::new()
     }
 
     /// Construct a [`Block`] from a [`Primitive`] value.